rename: 전체 directory 구조 수정
