+++ app/(main)/AiRecommendModal.module.css
