+++ app/(main)/page.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 import styles from './Main.module.css';
 import DefaultLayout from '@/app/components/DefaultLayout';
 import Link from 'next/link';
-import SearchAiSection from './components/SearchAiSection';
-import CategorySection from './components/CategorySection';
+import SearchAiSection from './SearchAiSection';
+import CategorySection from './CategorySection';
 import { getMeetings } from '@/lib/meetings';
 import { Meetings } from '@/types/meetings';
 import KakaoMap from '@/app/map/KakaoMap';
