+++ app/chat/ChatBadge.tsx
