+++ app/privacy/Privacy.module.css
@@ -1,0 +1,3 @@
+.privacy {
+  font-family: var(--sub-font-bold);
+}
