+++ app/privacy/page.tsx
@@ -1,0 +1,5 @@
+import styles from './Privacy.module.css';
+
+export default function Privacy() {
+  return <h1 className={styles.privacy}>Privacy</h1>;
+}
