+++ app/terms/Terms.module.css
@@ -1,0 +1,3 @@
+.terms {
+  font-family: var(--sub-font-bold);
+}
