+++ app/terms/page.tsx
@@ -1,0 +1,5 @@
+import styles from './Terms.module.css';
+
+export default function Terms() {
+  return <h1 className={styles.terms}>Terms</h1>;
+}
